+++ src/renderer/pages/ModsPage.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useState } from "react";
-import DividedSection from "../components/DividedSection";
 import MainPanel from "../components/MainPanel";
 import MinecraftButton from "../components/MinecraftButton";
-import { getMinecraftUWPFolder, getModsFolder } from "../versionSwitcher/AmethystPaths";
+import { MinecraftUWPFolder, ModsFolder } from "../scripts/Paths";
 
 import { validateModConfig, ModConfig } from "../scripts/Mods";
 
@@ -17,10 +16,9 @@ type ModErrorInfo = { modIdentifier: string, description?: string, modErrors: st
 function getAllMods(): ModErrorInfo[] {
     const results: ModErrorInfo[] = [];
 
-    const modsFolder = getModsFolder();
-    if (!fs.existsSync(modsFolder)) return results;
+    if (!fs.existsSync(ModsFolder)) return results;
 
-    const allModNames = fs.readdirSync(modsFolder, {withFileTypes: true})
+    const allModNames = fs.readdirSync(ModsFolder, {withFileTypes: true})
         .filter(f => f.isDirectory())
         .map(dir => dir.name);
 
@@ -38,7 +36,7 @@ function getAllMods(): ModErrorInfo[] {
         };
 
         // Validate that it has a config file
-        const modConfigPath = path.join(modsFolder, modIdentifier, "mod.json");
+        const modConfigPath = path.join(ModsFolder, modIdentifier, "mod.json");
 
         if (!fs.existsSync(modConfigPath)) {
             modErrors.push(`Missing mod.json configuration file inside mod folder.`);
@@ -68,16 +66,14 @@ function getAllMods(): ModErrorInfo[] {
 
 const openModsFolder = () => {
     // Don't reveal in explorer unless there is an existing minecraft folder
-    if (!fs.existsSync(getMinecraftUWPFolder())) {
+    if (!fs.existsSync(MinecraftUWPFolder)) {
         alert("Minecraft is not currently installed");
         return;
     }
 
-    const folder = getModsFolder();
+    if (!fs.existsSync(ModsFolder)) fs.mkdirSync(ModsFolder, {recursive: true});
 
-    if (!fs.existsSync(folder)) fs.mkdirSync(folder, {recursive: true});
-
-    const startGameCmd = `explorer "${folder}"`;
+    const startGameCmd = `explorer "${ModsFolder}"`;
     child.spawn(startGameCmd, {shell: true})
 }
  
@@ -95,58 +91,77 @@ export default function ModsPage() {
 
     return (
         <>
-        <MainPanel>
-            <div className="flex flex-col gap-[8px] h-full p-[8px] bg-[#48494A] border-[3px] border-[#1E1E1F]">
-                <p className="minecraft-seven text-white text-[14px]">Mod Manager</p>
-                <div className="h-full border-[3px] border-[#1E1E1F] bg-[#313233] overflow-y-auto scrollbar">
-                    { allReports.map((report) => 
-                        <div onClick={() => {setSelectedReport(report)}} key={report.modIdentifier}>
-                            <div className="cursor-pointer border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
-                                <p className="minecraft-seven text-white text-[14px] px-[4px]">{report.modIdentifier}</p>
-                                <p className="minecraft-seven text-[#B1B2B5] text-[14px] px-[4px]">{report.description}</p>
-                                {report.modErrors.length > 0 && (<p className="minecraft-seven text-red-400 text-[14px] px-[4px]">{report.modErrors.length} Errors!</p>)}
-                                {report.modErrors.length === 0 && <p className="minecraft-seven text-[#BCBEC0] text-[14px] px-[4px]">No Errors</p>}
+            <MainPanel>
+                <div className="flex flex-col gap-[8px] h-full p-[8px] bg-[#48494A] border-[3px] border-[#1E1E1F]">
+                    <p className="minecraft-seven text-white text-[14px]">Mod Manager</p>
+                    <div className="h-full border-[3px] border-[#1E1E1F] bg-[#313233] overflow-y-auto scrollbar">
+                        { allReports.map((report) =>
+                            <div onClick={() => {setSelectedReport(report)}} key={report.modIdentifier}>
+                                <div className="cursor-pointer border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+                                    <p className="minecraft-seven text-white text-[14px] px-[4px]">{report.modIdentifier}</p>
+                                    <p className="minecraft-seven text-[#B1B2B5] text-[14px] px-[4px]">{report.description}</p>
+                                    {report.modErrors.length > 0 && (<p className="minecraft-seven text-red-400 text-[14px] px-[4px]">{report.modErrors.length} Errors!</p>)}
+                                    {report.modErrors.length === 0 && <p className="minecraft-seven text-[#BCBEC0] text-[14px] px-[4px]">No Errors</p>}
+                                </div>
                             </div>
-                        </div>
-                    )}
-                </div>
-                <div className= "w-full h-fit">
-                    <MinecraftButton text="Open Mods Folder" onClick={openModsFolder}/>
-                </div>
-            </div>
-        </MainPanel>
-
-        {selectedReport && (<>
-            <div className="fixed top-0 left-0 w-full h-full bg-[#000000BB]" onClick={() => setSelectedReport(undefined)}></div>
-
-            <div className="fixed top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 flex flex-col items-center justify-center">
-                <DividedSection className="w-[500px]">
-                    <div className="flex">
-                        <p className="minecraft-seven text-white text-[12px] max-w-[400px]">{selectedReport.modIdentifier}</p>
-                        <p className="minecraft-seven text-[#BCBEC0] text-[12px] text-right ml-auto cursor-pointer" onClick={() => setSelectedReport(undefined)}>X</p>
+                        )}
                     </div>
+                    <div className= "w-full h-fit">
+                        <MinecraftButton text="Open Mods Folder" onClick={openModsFolder}/>
+                    </div>
+                </div>
+            </MainPanel>
 
-                    <p className="minecraft-seven text-[#BCBEC0] text-[12px] max-w-[400px]">{selectedReport.description ?? ""}</p>
-                </DividedSection>
-                {selectedReport.modErrors.length > 0 
-                    ? (
-                        <DividedSection className="w-[500px]">    
-                            <p className="minecraft-seven text-white text-[12px] pt-[6px]">Errors:</p>
-                            <ul>
-                                {selectedReport.modErrors.map(err => (
-                                    <li className="minecraft-seven text-red-400 text-[12px]" key={err}>- {err}</li>
-                                ))}
-                            </ul>
-                        </DividedSection>
-                    )
-                : (
-                    <DividedSection className="w-[500px]">
-                        <p className="minecraft-seven text-white text-[12px] pt-[6px]">No issues detected!</p>
-                    </DividedSection>
+            {
+                selectedReport &&
+
+                (
+                    <>
+                        <div className="fixed top-0 left-0 w-full h-full bg-[#000000BB]" onClick={() => setSelectedReport(undefined)}></div>
+
+                        <div className="fixed top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 flex flex-col items-center justify-center border-[3px] border-[#1E1E1F]">
+                            <div className="w-[500px] border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+                                <div className="flex">
+                                    <p className="minecraft-seven text-white text-[14px] max-w-[400px]">{selectedReport.modIdentifier}</p>
+                                    <div className="p-[4px] justify-center items-center ml-auto cursor-pointer" onClick={() => setSelectedReport(undefined)}>
+                                        <svg width="12" height="12" viewBox="0 0 12 12">
+                                            <polygon className="fill-[#FFFFFF]" fillRule="evenodd"
+                                                     points="11 1.576 6.583 6 11 10.424 10.424 11 6 6.583 1.576 11 1 10.424 5.417 6 1 1.576 1.576 1 6 5.417 10.424 1"/>
+                                        </svg>
+                                    </div>
+                                </div>
+
+                                <p className="minecraft-seven text-[#BCBEC0] text-[12px] max-w-[400px]">{selectedReport.description ?? ""}</p>
+                            </div>
+                            {
+                                selectedReport.modErrors.length > 0
+
+                                ?
+
+                                (
+                                    <div
+                                        className="w-[500px] border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+                                        <p className="minecraft-seven text-white text-[12px]">Errors:</p>
+                                        <ul>
+                                            {selectedReport.modErrors.map(err => (
+                                                <li className="minecraft-seven text-red-400 text-[12px]" key={err}>- {err}</li>
+                                            ))}
+                                        </ul>
+                                    </div>
+                                )
+
+                                :
+
+                                (
+                                    <div className="w-[500px] border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+                                        <p className="minecraft-seven text-white text-[12px]">No issues detected!</p>
+                                    </div>
+                                )
+                            }
+                        </div>
+                    </>
                 )
             }
-            </div>
-        </>)}
         </>
     )
 }
