+++ src/renderer/scripts/backend/Extractor.ts
