+++ src/renderer/scripts/backend/Progress.ts
