--- src/renderer/types/LauncherConfig.ts
@@ -1,8 +1,0 @@
-export interface LauncherConfig {
-    runtime: string,
-    mods: string[],
-    developer_mode: boolean,
-    keep_open: boolean,
-    selected_profile: number,
-    ui_theme: string
-}
