--- src/renderer/types/Profile.ts
@@ -1,6 +1,0 @@
-export type Profile = {
-    name: string,
-    runtime: string,
-    mods: string[]
-    minecraft_version: string
-};
